pyproject.toml
[tool.poetry]
name = "kotatsuneko-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
cryptography = "^42.0.8"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
python-jose = "^3.3.0"
firebase-admin = "^6.5.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
